refactor(server): extract Windows noise filter to utils/logging_filters.py
Task A of docs/current/server-py-module-split.md. Pure extraction,
verbatim move verified byte-identical against git history. Removed
now-dead top-level logging import from server.py (only consumer was
this block); functional check confirms the filter is still registered
on the asyncio logger from its new location.
